Added: implemented few methods of QuizDAO

+++ src/main/resources/sql/create_quest_tables.sql
@@ -9,7 +9,7 @@ CREATE TABLE quizzes (
 CREATE TABLE questions (
     id serial PRIMARY KEY,
     content text,
-    score integer,
+    score numeric,
     quiz integer REFERENCES quizzes(id)
 );
